modify page api: create van with vanType dto; set vanTypeName in response

+++ src/main/java/com/yumcamp/entity/Van.java
@@ -20,7 +20,7 @@ public class Van implements Serializable {
     @TableId(type = IdType.AUTO)
     private Long vanId;
     private String vanName;
-    private Integer vanTypeId;
+    private Long vanTypeId;
     private String vanLocation;
     private Integer berths;
     private String vanDescription;
